+++ src/main/java/com/redpxnda/respawnobelisks/RespawnObelisks.java
@@ -5,6 +5,7 @@ import com.redpxnda.respawnobelisks.config.ServerConfig;
 import com.redpxnda.respawnobelisks.event.RespawnObeliskEvents;
 import com.redpxnda.respawnobelisks.network.Packets;
 import com.redpxnda.respawnobelisks.registry.Registry;
+import net.minecraft.data.worldgen.Structures;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;